Update Assignment2.java
Creating the program to accept 5 employee IDs and the corresponding names and their salaries from the user and store them in three arrays. Passing these arrays to a function as arguments. This function display the content of the arrays. In addition, the same function display with Employee ID array and Employee name array as arguments by using concept of function overloading and it displays the content of the 2 arrays. Lastly, the function takes 4 arguments and search for the name in the Employee name array and displays its corresponding id and salary.